CUDA spiking neural net
This is an approximation of a spiking neural net using NVIDIA's CUDA
framework to make computations in parallel on the GPU.

It was made on a computer using Windows 7, an AMD Phenom(tm) II X4 965
processor and a NVIDIA GeForce GTX 460 graphics card. I was using
Microsoft Visual Studio this is the whole project file.